upload images and edit images

+++ vite-askfromcommunity/src/Components/ReportComponent/ReportPage.jsx
@@ -1,24 +1,51 @@
-import React, { useState } from 'react';
+import React, { useState, useEffect } from 'react';
 import ReportTable from './ReportTable';
 import './ReportPage.css'
 import ImageSlider from '../ImageSliderComponent/ImageSlider';
 import axios from 'axios';
 
-function ReportPage({item}) {
+function ReportPage({item, status}) {
   const [isDelete, setIsDelete] = useState(false);
   const [isRemove, setIsRemove] = useState(false);
+  const [posts, setPosts] = useState([]);
+
+  useEffect(() => {
+    if(status=="reply"){
+      getPost(item.post_id);
+    }
+  }, []);
 
   async function handleDelete(){
     setIsDelete(true);
-    alert("Post deleted!!!")
-    try{
-      await axios.delete(`http://localhost:80/project_1/AskFromCommunity/PostManager.php`,{
-      data:{
-        post_id:item.post_id,
-        status:"deletePost"
+      try{
+        if(status=="post"){          
+        alert("Post deleted!!!");
+        await axios.delete(`http://localhost:80/project_1/AskFromCommunity/PostManager.php`,{
+          data:{
+            post_id:item.post_id,
+            status:"deletePost"
+          }
+        });
+        }else{
+          alert("Reply deleted!!!");
+          await axios.delete(`http://localhost:80/project_1/AskFromCommunity/PostManager.php`,{
+          data:{
+            post_id:item.post_id,
+            status:"deletePost"
+          }
+        });
+        }
+        
+      }catch(err){
+        console.error(err);
       }
-    });
-    }catch(err){
+  }
+    
+  async function getPost($post_id) {
+    try {
+      const res = await axios.get(`http://localhost:80/project_1/AskFromCommunity/PostManager.php?post_id=${$post_id}`);
+      setPosts(res.data);
+    } catch (err) {
       console.error(err);
     }
   }
@@ -27,17 +54,25 @@ function ReportPage({item}) {
     setIsRemove(true);
     alert("Romoved All Reports!!!")
     try{
-      await axios.delete(`http://localhost:80/project_1/AskFromCommunity/PostManager.php`,{
-      data:{
-        post_id:item.post_id,
-        status:"removeReports"
+      if(status=="post"){
+        await axios.delete(`http://localhost:80/project_1/AskFromCommunity/PostManager.php`,{
+          data:{
+            post_id:item.post_id,
+            status:"removeReports"
+          }
+        });
+      }else{
+        await axios.delete(`http://localhost:80/project_1/AskFromCommunity/PostManager.php`,{
+          data:{
+            post_id:item.post_id,
+            status:"removeReports"
+          }
+        });
       }
-    });
     }catch(err){
       console.error(err);
     }
   }
-
   return (
     <>
       <div className='report-page-container'>
@@ -45,16 +80,23 @@ function ReportPage({item}) {
           <button id='rejectBtn' onClick={handleRemoveReports} disabled={isRemove}>Reject All Reports</button>
           <button id='deleteBtn' onClick={handleDelete} disabled={isDelete}>Delete Post</button>
           <div id='imageSlider-container'>
-            <ImageSlider />
+          <ImageSlider post_id={item.post_id}/>
           </div>
-          
-          <h2>{item.title}</h2>
-          <p>{item.description}</p>
+          {(status=="post"?true:false)&&<h2>{item.title}</h2>}
+          {(status=="post"?true:false)&&<p>{item.description}</p>}
+          {(status=="reply"?true:false)&& posts.map((item) => (
+            <div key ={item.post_id}>
+              <h2>{item.title}</h2>
+              <p>{item.description}</p>
+            </div >            
+          ))}
         </div>
         <hr />
         <br />
         <div className='report-details-container'>
-        <ReportTable post_id={item.post_id} />
+        {(status=="post"?true:false)&&<ReportTable status={status} post_id={item.post_id} reply_id={""} />}
+
+        {(status=="reply"?true:false)&&<ReportTable status={status} post_id={item.post_id} reply_id={item.reply_id} />}
         </div>
       </div>
     </>
